+++ MSAL/src/requests/MSALSilentRequest.h
@@ -30,7 +30,7 @@
 @interface MSALSilentRequest : MSALBaseRequest
 {
     BOOL _forceRefresh;
-    uint _expirationBuffer;
+    NSUInteger _expirationBuffer;
 }
 
 @property NSString *state;
@@ -38,7 +38,7 @@
 - (id)initWithParameters:(MSALRequestParameters *)parameters
             forceRefresh:(BOOL)forceRefresh
               tokenCache:(MSIDDefaultTokenCacheAccessor *)tokenCache
-        expirationBuffer:(uint)expirationBuffer
+        expirationBuffer:(NSUInteger)expirationBuffer
                    error:(NSError *__autoreleasing *)error;
 
 @end
